+++ frontend/app/app.component.ts
@@ -2,7 +2,7 @@ import { Component } from "@angular/core";
 import { ROUTER_DIRECTIVES } from "@angular/router";
 
 // Add the RxJS Observable operators we need in this app.
-import "../rxjs-operators";
+import "./shared/rxjs-operators";
 
 @Component({
     moduleId: module.id,
