+++ frontend/app/config/sensitive_conf_dummy.ts
@@ -1,0 +1,6 @@
+/**
+ * Class for the sensitive config, which overwrites values given in the production config.
+ * It is changed in the production environment, else never used.
+ */
+export class SensitiveConfig {
+}
