+++ frontend/app/events/event.component.scss
