+++ frontend/app/events/event.component.spec.ts
