+++ frontend/app/events/event.component.ts
