+++ frontend/app/map/map.component.scss
