+++ frontend/app/map/map.component.spec.ts
