+++ frontend/app/map/map.component.ts
