+++ frontend/app/map/mapObject.ts
