+++ frontend/app/shared/rxjs-operators.ts
