+++ frontend/main.ts
@@ -3,7 +3,7 @@
 
 import { platformBrowserDynamic } from "@angular/platform-browser-dynamic";
 
-import { Config } from "./config/config";
+import { Config } from "./app/config/config";
 import { AppModule } from "./app/app.module";
 
 let config = new Config();
